requirements.txt
tensorflow-gpu
keras
geopandas==0.9.0
networkx==2.5.1
igraph==0.9.8
scipy==1.5.4
numpy==1.19.5
ujson==4.2.0
bs4==0.0.1
pandas==1.1.5
shapley==1.0.2
fiona==1.8.20
haversine==2.5.1
geographiclib==1.52
